Refactor app.py and utils/adm.py to add login functionality and restricted area
