+++ app/app.py
@@ -219,3 +219,5 @@ if dev_data:
 
 else:
     st.error("Não foi possível recuperar os dados do usuário.")
+
+
