+++ app/config/variaveis_globais.py
@@ -15,8 +15,12 @@ mec_image = os.path.join(os.path.dirname(__file__), '../../images/mec.png')
 
 
 arquivo_de_apresentacao = os.path.join(os.path.dirname(__file__), '../../data/interim/apresentacao.txt')
-arquivo_de_teste = os.path.join(os.path.dirname(__file__), '../../data/interim/teste.txt')
-arquivo_de_rubrica = os.path.join(os.path.dirname(__file__), '../../data/interim/rubricas.txt')
+arquivo_de_teste_tp2 = os.path.join(os.path.dirname(__file__), '../../data/interim/teste_tp2.txt')
+arquivo_de_rubrica_tp2 = os.path.join(os.path.dirname(__file__), '../../data/interim/rubricas_tp2.txt')
+
+arquivo_de_teste_tp3= os.path.join(os.path.dirname(__file__), '../../data/interim/teste_tp3.txt')
+arquivo_de_rubrica_tp3 = os.path.join(os.path.dirname(__file__), '../../data/interim/rubricas_tp3.txt')
+
 arquivo_de_palavras = os.path.join(os.path.dirname(__file__), '../../data/external/dados_completos.json')
 
 
@@ -25,6 +29,12 @@ arquivo_de_resposta1 = os.path.join(os.path.dirname(__file__), '../../data/inter
 arquivo_de_resposta2 = os.path.join(os.path.dirname(__file__), '../../data/interim/answers/resposta2.txt')
 arquivo_de_resposta3 = os.path.join(os.path.dirname(__file__), '../../data/interim/answers/resposta3.txt')
 arquivo_de_resposta4 = os.path.join(os.path.dirname(__file__), '../../data/interim/answers/resposta4.txt')
+
+arquivo_de_respostatp3_1 = os.path.join(os.path.dirname(__file__), '../../data/interim/answers/respostatp3_1.txt')
+arquivo_de_respostatp3_2 = os.path.join(os.path.dirname(__file__), '../../data/interim/answers/respostatp3_2.txt')
+arquivo_de_respostatp3_3 = os.path.join(os.path.dirname(__file__), '../../data/interim/answers/respostatp3_3.txt')
+arquivo_de_respostatp3_4 = os.path.join(os.path.dirname(__file__), '../../data/interim/answers/respostatp3_4.txt')
+arquivo_de_respostatp3_5 = os.path.join(os.path.dirname(__file__), '../../data/interim/answers/respostatp3_5.txt')
 
 config_vars = create_global_variables(streamlit_secret)
