+++ app/utils/render.py
@@ -9,24 +9,14 @@ from utils.database import update_document, get_collection
 from config.variaveis_globais import (
     streamlit_secret, 
     image_directory,  
-    arquivo_de_apresentacao, 
-    arquivo_de_teste, 
-    arquivo_de_rubrica
 )
 
 
-from utils.abas import (adicionar_conteudo,
-                        exibir_respostas,
-                        exibir_pacotes_viagem,
-                        exibir_area_restrita,
-                        exibir_api_teste,)
-
-from utils.markdown import read_markdown_file
 from utils.background import get_cached_random_image
 from utils.globals import create_global_variables
 
 from utils.frescuras import (contar_itens_config)
-
+from utils.tab_contents import tab_contents
 
 
 #################################################################################
@@ -46,43 +36,9 @@ def render_tabs(selected_section, menu_dados):
     num_sub_tabs_test = contar_itens_config(config_vars, 'sub_tab_test') + 1
 
     # Criar abas principais
-    main_tab_names = [config_vars[f'main_tab_{i}'] for i in range(1, num_main_tabs)] 
+    main_tab_names = [config_vars[f'main_tab_{i}'] for i in range(1, num_main_tabs)]
 
-    def tab_contents():
-
-        main_tabs = st.tabs(main_tab_names)
-
-        for i, tab in enumerate(main_tabs):
-            with tab:
-                if main_tab_names[i] == config_vars['main_tab_2']:  # "O Projeto"
-                    adicionar_conteudo(lambda: read_markdown_file(arquivo_de_apresentacao))
-                
-                elif main_tab_names[i] == config_vars['main_tab_3']:  # "O Teste"
-                    sub_tab_names = [config_vars[f'sub_tab_test_{j}'] for j in range(1, num_sub_tabs_test)]
-                    sub_tabs = st.tabs(sub_tab_names)
-                    
-                    for j, sub_tab in enumerate(sub_tabs):
-                        with sub_tab:
-                            if j == 0:
-                                adicionar_conteudo(lambda: read_markdown_file(arquivo_de_teste))
-                            elif j == 1:
-                                adicionar_conteudo(lambda: read_markdown_file(arquivo_de_rubrica))
-                            elif j == 2:
-                                adicionar_conteudo(lambda: exibir_respostas(selected_section, menu_dados))
-                
-                elif main_tab_names[i] == config_vars['main_tab_1']:  # "O App"
-                    sub_tab_names = [config_vars[f'sub_tab_{j}'] for j in range(1, num_sub_tabs)]
-                    sub_tabs = st.tabs(sub_tab_names)
-                    
-                    for j, sub_tab in enumerate(sub_tabs):
-                        with sub_tab:
-                            if j == 0:
-                                adicionar_conteudo(exibir_pacotes_viagem)
-                            elif j == 1:
-                                adicionar_conteudo(exibir_area_restrita)
-                            elif j == 2:
-                                adicionar_conteudo(exibir_api_teste)
-    return tab_contents
+    return lambda: tab_contents(main_tab_names, num_sub_tabs, num_sub_tabs_test, selected_section, menu_dados)
 
 def render_alterar_senha_form():
     st.markdown("##### Alterar Senha")
@@ -114,3 +70,18 @@ def render_alterar_senha_form():
                     st.success("Senha alterada com sucesso!")
                 else:
                     st.error("Senha atual incorreta.")
+
+
+def adicionar_conteudo(conteudo, *args, **kwargs):
+    try:
+        if isinstance(conteudo, str):
+            st.markdown(conteudo)
+        elif callable(conteudo):
+            result = conteudo(*args, **kwargs)
+            if result is not None:
+                st.write(result)
+        else:
+            st.write(conteudo)
+    except Exception as e:
+        st.error(f"Erro ao adicionar conteúdo: {str(e)}")
+        print(f"Erro ao adicionar conteúdo: {str(e)}")
